properties: split out editor GUI
